requirements.txt
streamlit==1.39.0
cryptography==42.0.0
Pillow==10.3.0
stegano==0.11.3
plotly==5.18.0
pandas==2.2.0
python-magic-bin==0.4.14
streamlit-option-menu==0.3.12
extra-streamlit-components==0.1.60
